+++ jib-gradle-plugin/src/main/java/com/google/cloud/tools/jib/gradle/DockerContextTask.java
@@ -85,11 +85,11 @@ public class DockerContextTask extends DefaultTask {
   }
 
   /**
-   * The output directory can be overriden with the {@code --targetDir} command line option.
+   * The output directory can be overriden with the {@code --jibTargetDir} command line option.
    *
    * @param targetDir the output directory.
    */
-  @Option(option = "targetDir", description = "Directory to output the Docker context to")
+  @Option(option = "jibTargetDir", description = "Directory to output the Docker context to")
   public void setTargetDir(String targetDir) {
     this.targetDir = targetDir;
   }
@@ -134,7 +134,7 @@ public class DockerContextTask extends DefaultTask {
     } catch (IOException ex) {
       throw new GradleException(
           HelpfulSuggestionsProvider.get("Export Docker context failed")
-              .suggest("check if the command-line option `--targetDir` is set correctly"),
+              .suggest("check if the command-line option `--jibTargetDir` is set correctly"),
           ex);
     }
   }