+++ jib-maven-plugin/src/main/java/com/google/cloud/tools/jib/maven/DockerContextMojo.java
@@ -40,7 +40,7 @@ public class DockerContextMojo extends JibPluginConfiguration {
 
   @Nullable
   @Parameter(
-      property = "jib.targetDir",
+      property = "jibTargetDir",
       defaultValue = "${project.build.directory}/jib-docker-context",
       required = true)
   private String targetDir;