+++ jib-maven-plugin/CHANGELOG.md
@@ -12,7 +12,7 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
-- Docker context export parameter `-Djib.dockerDir` to `-Djib.targetDir` for consistency ([#662](https://github.com/GoogleContainerTools/jib/issues/662))
+- Docker context export parameter `-Djib.dockerDir` to `-DjibTargetDir` ([#662](https://github.com/GoogleContainerTools/jib/issues/662))
 
 ### Fixed